requirements.txt
# Dependencias del proyecto
psycopg2-binary>=2.9.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytz>=2023.3

# PostgreSQL debe estar instalado y corriendo
# Configura la conexión en config/settings.py

